+++ .appveyor.yml
@@ -20,11 +20,6 @@ environment:
   matrix:
   - db: mssql
     driver: sqlsrv
-    db_version: sql2008r2sp2
-    coverage: yes
-    php: 7.3
-  - db: mssql
-    driver: sqlsrv
     db_version: sql2012sp1
     php: 7.3
     coverage: yes
