+++ docs/en/reference/platforms.rst
@@ -50,8 +50,7 @@ Oracle
 Microsoft SQL Server
 ^^^^^^^^^^^^^^^^^^^^
 
--  ``SQLServerPlatform`` for version 2008 and above.
--  ``SQLServer2012Platform`` for version 2012 and above.
+-  ``SQLServerPlatform`` for version 2012 and above.
 
 PostgreSQL
 ^^^^^^^^^^
