+++ lib/Doctrine/DBAL/Driver/AbstractSQLServerDriver.php
@@ -5,51 +5,17 @@ declare(strict_types=1);
 namespace Doctrine\DBAL\Driver;
 
 use Doctrine\DBAL\Connection;
+use Doctrine\DBAL\Driver;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
-use Doctrine\DBAL\Platforms\Exception\InvalidPlatformVersion;
-use Doctrine\DBAL\Platforms\SQLServer2012Platform;
 use Doctrine\DBAL\Platforms\SQLServerPlatform;
 use Doctrine\DBAL\Schema\AbstractSchemaManager;
 use Doctrine\DBAL\Schema\SQLServerSchemaManager;
-use Doctrine\DBAL\VersionAwarePlatformDriver;
-use function preg_match;
-use function version_compare;
 
 /**
  * Abstract base implementation of the {@link Doctrine\DBAL\Driver} interface for Microsoft SQL Server based drivers.
  */
-abstract class AbstractSQLServerDriver implements VersionAwarePlatformDriver
+abstract class AbstractSQLServerDriver implements Driver
 {
-    /**
-     * {@inheritdoc}
-     */
-    public function createDatabasePlatformForVersion(string $version) : AbstractPlatform
-    {
-        if (! preg_match(
-            '/^(?P<major>\d+)(?:\.(?P<minor>\d+)(?:\.(?P<patch>\d+)(?:\.(?P<build>\d+))?)?)?/',
-            $version,
-            $versionParts
-        )) {
-            throw InvalidPlatformVersion::new(
-                $version,
-                '<major_version>.<minor_version>.<patch_version>.<build_version>'
-            );
-        }
-
-        $majorVersion = $versionParts['major'];
-        $minorVersion = $versionParts['minor'] ?? 0;
-        $patchVersion = $versionParts['patch'] ?? 0;
-        $buildVersion = $versionParts['build'] ?? 0;
-        $version      = $majorVersion . '.' . $minorVersion . '.' . $patchVersion . '.' . $buildVersion;
-
-        switch (true) {
-            case version_compare($version, '11.00.2100', '>='):
-                return new SQLServer2012Platform();
-            default:
-                return new SQLServerPlatform();
-        }
-    }
-
     /**
      * {@inheritdoc}
      */
