+++ lib/Doctrine/DBAL/Platforms/Keywords/SQLServerKeywords.php
@@ -173,6 +173,9 @@ class SQLServerKeywords extends KeywordList
             'SCHEMA',
             'SECURITYAUDIT',
             'SELECT',
+            'SEMANTICKEYPHRASETABLE',
+            'SEMANTICSIMILARITYDETAILSTABLE',
+            'SEMANTICSIMILARITYTABLE',
             'SESSION_USER',
             'SET',
             'SETUSER',
@@ -190,6 +193,7 @@ class SQLServerKeywords extends KeywordList
             'TRANSACTION',
             'TRIGGER',
             'TRUNCATE',
+            'TRY_CONVERT',
             'TSEQUAL',
             'UNION',
             'UNIQUE',
@@ -206,6 +210,7 @@ class SQLServerKeywords extends KeywordList
             'WHERE',
             'WHILE',
             'WITH',
+            'WITHIN GROUP',
             'WRITETEXT',
         ];
     }
