+++ tests/Doctrine/Tests/DBAL/Driver/AbstractSQLServerDriverTest.php
@@ -8,7 +8,6 @@ use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Driver;
 use Doctrine\DBAL\Driver\AbstractSQLServerDriver;
 use Doctrine\DBAL\Platforms\AbstractPlatform;
-use Doctrine\DBAL\Platforms\SQLServer2012Platform;
 use Doctrine\DBAL\Platforms\SQLServerPlatform;
 use Doctrine\DBAL\Schema\AbstractSchemaManager;
 use Doctrine\DBAL\Schema\SQLServerSchemaManager;
@@ -36,24 +35,7 @@ class AbstractSQLServerDriverTest extends AbstractDriverTest
     protected function getDatabasePlatformsForVersions() : array
     {
         return [
-            ['10', SQLServerPlatform::class],
-            ['10.00', SQLServerPlatform::class],
-            ['10.00.0', SQLServerPlatform::class],
-            ['10.00.1599', SQLServerPlatform::class],
-            ['10.00.1599.99', SQLServerPlatform::class],
-            ['10.00.1600', SQLServerPlatform::class],
-            ['10.00.1600.0', SQLServerPlatform::class],
-            ['10.00.1600.99', SQLServerPlatform::class],
-            ['10.00.1601', SQLServerPlatform::class],
-            ['10.10', SQLServerPlatform::class],
-            ['10.10.9999', SQLServerPlatform::class],
-            ['11.00.2099', SQLServerPlatform::class],
-            ['11.00.2099.99', SQLServerPlatform::class],
-            ['11.00.2100', SQLServer2012Platform::class],
-            ['11.00.2100.0', SQLServer2012Platform::class],
-            ['11.00.2100.99', SQLServer2012Platform::class],
-            ['11.00.2101', SQLServer2012Platform::class],
-            ['12', SQLServer2012Platform::class],
+            ['12', SQLServerPlatform::class],
         ];
     }
 }
